# add channels last support for ReplicationPad on CPU