--- a/aten/src/ATen/native/cpu/PaddingKernel.cpp
+++ b/aten/src/ATen/native/cpu/PaddingKernel.cpp
@@ -81,6 +81,20 @@ struct ReflectionPad {
   }
 };
 
+struct ReplicationPad {
+  static int64_t index(int64_t j, int64_t size, int64_t pad, int64_t offset) {
+    int64_t i;
+    if (j < pad) {
+      i = pad;
+    } else if (j >= pad && j < size + pad) {
+      i = j;
+    } else {
+      i = size + pad - 1;
+    }
+    return i + offset;
+  }
+};
+
 template <typename scalar_t>
 static inline void copy_stub(scalar_t* out, const scalar_t* in, int64_t size) {
   using Vec = Vectorized<scalar_t>;
@@ -580,6 +594,104 @@ void reflection_pad3d_backward_kernel_impl(
   }
 }
 
+// replication padding
+void replication_pad1d_kernel_impl(const Tensor& output, const Tensor& input, IntArrayRef padding) {
+  PaddingParams param{input, output, padding};
+  AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(input.scalar_type(), "replication_pad1d", [&] {
+    cpu_padding<scalar_t, ReplicationPad>(output, input, param);
+  });
+}
+
+void replication_pad1d_backward_kernel_impl(
+    const Tensor& grad_input, const Tensor& grad_output, IntArrayRef padding) {
+  PaddingParams param{grad_input, grad_output, padding};
+  AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(grad_output.scalar_type(), "replication_pad1d_backward", [&] {
+    cpu_padding_backward<scalar_t, ReplicationPad>(grad_input, grad_output, param);
+  });
+}
+
+void replication_pad2d_kernel_impl(const Tensor& output, const Tensor& input, IntArrayRef padding) {
+  PaddingParams param{input, output, padding};
+  switch (input.suggest_memory_format()) {
+    case at::MemoryFormat::Contiguous: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(input.scalar_type(), "replication_pad2d", [&] {
+        cpu_padding<scalar_t, ReplicationPad>(output, input, param);
+      });
+      break;
+    }
+    case at::MemoryFormat::ChannelsLast: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(input.scalar_type(), "replication_pad2d_channels_last", [&]{
+        cpu_padding_channels_last<scalar_t, ReplicationPad>(output, input, param);
+      });
+      break;
+    }
+    default:
+      TORCH_CHECK(false, "Unsupported memory format. Supports only ChannelsLast, Contiguous");
+  }
+}
+
+void replication_pad2d_backward_kernel_impl(
+    const Tensor& grad_input, const Tensor& grad_output, IntArrayRef padding) {
+  PaddingParams param{grad_input, grad_output, padding};
+  switch (grad_output.suggest_memory_format()) {
+    case at::MemoryFormat::Contiguous: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(grad_output.scalar_type(), "replication_pad2d_backward", [&] {
+        cpu_padding_backward<scalar_t, ReplicationPad>(grad_input, grad_output, param);
+      });
+      break;
+    }
+    case at::MemoryFormat::ChannelsLast: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(grad_output.scalar_type(), "replication_pad2d_backward_channels_last", [&]{
+        cpu_padding_backward_channels_last<scalar_t, ReplicationPad>(grad_input, grad_output, param);
+      });
+      break;
+    }
+    default:
+      TORCH_CHECK(false, "Unsupported memory format. Supports only ChannelsLast, Contiguous");
+  }
+}
+
+void replication_pad3d_kernel_impl(const Tensor& output, const Tensor& input, IntArrayRef padding) {
+  PaddingParams param{input, output, padding};
+  switch (padding_memory_format_3d(input)) {
+    case at::MemoryFormat::Contiguous: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(input.scalar_type(), "replication_pad3d", [&] {
+        cpu_padding<scalar_t, ReplicationPad>(output, input, param);
+      });
+      break;
+    }
+    case at::MemoryFormat::ChannelsLast3d: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(input.scalar_type(), "replication_pad3d_channels_last", [&]{
+        cpu_padding_channels_last<scalar_t, ReplicationPad>(output, input, param);
+      });
+      break;
+    }
+    default:
+      TORCH_CHECK(false, "Unsupported memory format. Supports only ChannelsLast3d, Contiguous");
+  }
+}
+
+void replication_pad3d_backward_kernel_impl(
+    const Tensor& grad_input, const Tensor& grad_output, IntArrayRef padding) {
+  PaddingParams param{grad_input, grad_output, padding};
+  switch (padding_memory_format_3d(grad_output)) {
+    case at::MemoryFormat::Contiguous: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(grad_output.scalar_type(), "replication_pad3d_backward", [&] {
+        cpu_padding_backward<scalar_t, ReplicationPad>(grad_input, grad_output, param);
+      });
+      break;
+    }
+    case at::MemoryFormat::ChannelsLast3d: {
+      AT_DISPATCH_FLOATING_AND_COMPLEX_TYPES(grad_output.scalar_type(), "replication_pad3d_backward_channels_last", [&]{
+        cpu_padding_backward_channels_last<scalar_t, ReplicationPad>(grad_input, grad_output, param);
+      });
+      break;
+    }
+    default:
+      TORCH_CHECK(false, "Unsupported memory format. Supports only ChannelsLast3d, Contiguous");
+  }
+}
+
 } // anonymous namespace
 
 // reflection padding
@@ -589,5 +701,13 @@ REGISTER_DISPATCH(reflection_pad2d_kernel, &reflection_pad2d_kernel_impl);
 REGISTER_DISPATCH(reflection_pad2d_backward_kernel, &reflection_pad2d_backward_kernel_impl);
 REGISTER_DISPATCH(reflection_pad3d_kernel, &reflection_pad3d_kernel_impl);
 REGISTER_DISPATCH(reflection_pad3d_backward_kernel, &reflection_pad3d_backward_kernel_impl);
+
+// replication padding
+REGISTER_DISPATCH(replication_pad1d_kernel, &replication_pad1d_kernel_impl);
+REGISTER_DISPATCH(replication_pad1d_backward_kernel, &replication_pad1d_backward_kernel_impl);
+REGISTER_DISPATCH(replication_pad2d_kernel, &replication_pad2d_kernel_impl);
+REGISTER_DISPATCH(replication_pad2d_backward_kernel, &replication_pad2d_backward_kernel_impl);
+REGISTER_DISPATCH(replication_pad3d_kernel, &replication_pad3d_kernel_impl);
+REGISTER_DISPATCH(replication_pad3d_backward_kernel, &replication_pad3d_backward_kernel_impl);
 
 } // at::native